# public/api-docs/swagger-spec.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/beneficio/{id}": {
      "get": {
        "description": "Retorna os detalhes completos de um tipo de benefício específico.",
        "operationId": "Beneficio_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Benefício encontrado com sucesso",
            "content": {
              "application/json": {
                "example": ""
              }
            }
          },
          "404": {
            "description": "Benefício não encontrado",
            "content": {
              "application/json": {
                "example": ""
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter detalhes de um benefício",
        "tags": [
          "Benefícios"
        ]
      },
      "put": {
        "description": "Atualiza os dados de um tipo de benefício existente.",
        "operationId": "Beneficio_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTipoBeneficioDto"
              },
              "examples": {
                "Atualização de Benefício": {
                  "value": ""
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Benefício atualizado com sucesso",
            "content": {
              "application/json": {
                "example": ""
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "example": ""
              }
            }
          },
          "404": {
            "description": "Benefício não encontrado",
            "content": {
              "application/json": {
                "example": ""
              }
            }
          },
          "409": {
            "description": "Nome já em uso",
            "content": {
              "application/json": {
                "example": ""
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar tipo de benefício existente",
        "tags": [
          "Benefícios"
        ]
      }
    },
...
    "/api/v1/solicitacao": {
      "post": {
        "operationId": "Solicitacao_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateSolicitacaoDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Solicitação criada com sucesso"
          },
          "400": {
            "description": "Dados inválidos"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Criar nova solicitação de benefício",
        "tags": [
          "Solicitações"
        ]
      },
      "get": {
        "operationId": "Solicitacao_findAll",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "description": "Página atual",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "limit",
            "required": false,
            "in": "query",
            "description": "Itens por página",
            "schema": {
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "description": "Filtro por status",
            "schema": {
              "enum": [
                "rascunho",
                "pendente",
                "em_analise",
                "aguardando_documentos",
                "aprovada",
                "reprovada",
                "liberada",
                "cancelada"
              ],
              "type": "string"
            }
          },
          {
            "name": "unidade_id",
            "required": false,
            "in": "query",
            "description": "Filtro por unidade",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "beneficio_id",
            "required": false,
            "in": "query",
            "description": "Filtro por tipo de benefício",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "protocolo",
            "required": false,
            "in": "query",
            "description": "Busca por protocolo",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "data_inicio",
            "required": false,
            "in": "query",
            "description": "Data inicial (formato: YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "data_fim",
            "required": false,
            "in": "query",
            "description": "Data final (formato: YYYY-MM-DD)",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de solicitações retornada com sucesso"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar solicitações",
        "tags": [
          "Solicitações"
        ]
      }
    },
    "/api/v1/solicitacao/{id}": {
      "get": {
        "operationId": "Solicitacao_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Solicitação encontrada com sucesso"
          },
          "404": {
            "description": "Solicitação não encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Obter detalhes de uma solicitação",
        "tags": [
          "Solicitações"
        ]
      },
      "put": {
        "operationId": "Solicitacao_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "description": "Dados da solicitação atualizada",
          "content": {
            "application/json": {
              "schema": {
                "example": {
                  "nome": "João Silva",
                  "cpf": "12345678901",
                  "data_nascimento": "1990-01-01",
                  "endereco": {
                    "logradouro": "Rua Exemplo",
                    "numero": "123",
                    "bairro": "Bairro Exemplo",
                    "cidade": "Cidade Exemplo",
                    "estado": "Estado Exemplo",
                    "cep": "12345678"
                  },
                  "contato": {
                    "telefone": "123456789",
                    "email": "joao.silva@example.com"
                  },
                  "beneficio": {
                    "id": 1,
                    "nome": "Benefício Exemplo"
                  },
                  "status": "EM_ANALISE"
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Solicitação atualizada com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "example": {
                    "id": 1,
                    "nome": "João Silva",
                    "cpf": "12345678901",
                    "data_nascimento": "1990-01-01",
                    "endereco": {
                      "logradouro": "Rua Exemplo",
                      "numero": "123",
                      "bairro": "Bairro Exemplo",
                      "cidade": "Cidade Exemplo",
                      "estado": "Estado Exemplo",
                      "cep": "12345678"
                    },
                    "contato": {
                      "telefone": "123456789",
                      "email": "joao.silva@example.com"
                    },
                    "beneficio": {
                      "id": 1,
                      "nome": "Benefício Exemplo"
                    },
                    "status": "EM_ANALISE"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos"
          },
          "404": {
            "description": "Solicitação não encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar solicitação existente",
        "tags": [
          "Solicitações"
        ]
      }
    },
    "/api/v1/solicitacao/{id}/status": {
      "patch": {
        "description": "Atualiza o status de uma solicitação de benefício e registra o histórico da alteração.",
        "operationId": "SolicitacaoBeneficio_updateStatus",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateStatusSolicitacaoDto"
              },
              "examples": {
                "Atualização de Status": {}
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Status atualizado com sucesso",
            "content": {
              "application/json": {
                "example": {
                  "message": "Status atualizado com sucesso"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos ou transição de status inválida",
            "content": {
              "application/json": {
                "example": ""
              }
            }
          },
          "404": {
            "description": "Solicitação não encontrada",
            "content": {
              "application/json": {
                "example": ""
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Atualizar status de uma solicitação de benefício",
        "tags": [
          "Benefícios"
        ]
      }
    },
    "/api/v1/solicitacao/{id}/historico": {
      "get": {
        "operationId": "Solicitacao_getHistorico",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Histórico retornado com sucesso"
          },
          "404": {
            "description": "Solicitação não encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar histórico de uma solicitação",
        "tags": [
          "Solicitações"
        ]
      }
    },
...
    "/api/v1/documento/solicitacao/{solicitacaoId}": {
      "get": {
        "operationId": "Documento_findBySolicitacao",
        "parameters": [
          {
            "name": "solicitacaoId",
            "required": true,
            "in": "path",
            "description": "ID da solicitação",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "tipo",
            "required": false,
            "in": "query",
            "description": "Filtrar por tipo de documento",
            "schema": {
              "enum": [
                "COMPROVANTE_RESIDENCIA",
                "COMPROVANTE_RENDA",
                "RG",
                "CPF",
                "CARTAO_NIS"
              ],
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Lista de documentos retornada com sucesso"
          },
          "401": {
            "description": "Não autorizado"
          },
          "404": {
            "description": "Solicitação não encontrada"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "summary": "Listar documentos de uma solicitação",
        "tags": [
          "Documentos"
        ]
      }
    },
...
  },
  "info": {
    "title": "API de Gestão de Benefícios Eventuais",
...
    "version": "1.0",
    "contact": {
      "name": "Equipe SEMTAS",
      "url": "https://www.natal.rn.gov.br/semtas",
      "email": "suporte@semtas.natal.rn.gov.br"
    },
    "license": {
      "name": "Uso Interno",
      "url": "https://www.natal.rn.gov.br/semtas/licenca"
    }
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "JWT-auth": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "name": "JWT",
        "description": "Insira o token JWT",
        "in": "header"
      }
    },
    "schemas": {
...
      "CriteriosElegibilidadeDto": {
        "type": "object",
        "properties": {
          "idade_minima": {
            "type": "number",
            "description": "Idade mínima para elegibilidade"
          },
          "idade_maxima": {
            "type": "number",
            "description": "Idade máxima para elegibilidade"
          },
          "renda_maxima": {
            "type": "number",
            "description": "Renda máxima per capita para elegibilidade"
          },
          "tempo_minimo_residencia": {
            "type": "number",
            "description": "Tempo mínimo de residência no município (em meses)"
          },
          "outros": {
            "description": "Outros critérios de elegibilidade",
            "type": "array",
            "items": {
              "type": "string"
            }
          }
        }
      },
...
      "UpdateTipoBeneficioDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "description": "Nome do tipo de benefício",
            "example": "Auxílio Moradia"
          },
          "descricao": {
            "type": "string",
            "description": "Descrição detalhada do benefício"
          },
          "periodicidade": {
            "type": "string",
            "description": "Periodicidade do benefício",
            "enum": [
              "unico",
              "mensal",
              "trimestral",
              "semestral",
              "anual"
            ],
            "example": "mensal"
          },
          "base_juridica": {
            "type": "string",
            "description": "Base jurídica que fundamenta o benefício"
          },
          "valor": {
            "type": "number",
            "description": "Valor do benefício",
            "example": 400
          },
          "ativo": {
            "type": "boolean",
            "description": "Status do benefício",
            "default": true
          },
          "criterios_elegibilidade": {
            "description": "Critérios de elegibilidade para o benefício",
            "allOf": [
              {
                "$ref": "#/components/schemas/CriteriosElegibilidadeDto"
              }
            ]
          }
        }
      },
...
      "UpdateStatusSolicitacaoDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "description": "Novo status da solicitação",
            "example": "APROVADA",
            "enum": [
              "PENDENTE",
              "ANALISE",
              "APROVADA",
              "REJEITADA",
              "CANCELADA",
              "EM_PROCESSAMENTO",
              "CONCLUIDA"
            ]
          },
          "justificativa": {
            "type": "string",
            "description": "Justificativa para a mudança de status",
            "example": "Documentação completa e requisitos atendidos"
          }
        },
        "required": [
          "status"
        ]
      },
      "DocumentoSolicitacaoDto": {
        "type": "object",
        "properties": {
          "nome": {
            "type": "string",
            "description": "Nome do documento",
            "example": "Comprovante de Residência"
          },
          "tipo": {
            "type": "string",
            "description": "Tipo do documento",
            "example": "comprovante_residencia"
          },
          "arquivo_url": {
            "type": "string",
            "description": "URL ou caminho do arquivo",
            "example": "/uploads/documentos/12345.pdf"
          },
          "observacoes": {
            "type": "string",
            "description": "Observações sobre o documento"
          }
        },
        "required": [
          "nome",
          "tipo",
          "arquivo_url"
        ]
      },
      "CreateSolicitacaoDto": {
        "type": "object",
        "properties": {
          "beneficiario_id": {
            "type": "string",
            "description": "ID do cidadão beneficiário"
          },
          "tipo_beneficio_id": {
            "type": "string",
            "description": "ID do tipo de benefício solicitado"
          },
          "unidade_id": {
            "type": "string",
            "description": "ID da unidade onde a solicitação está sendo feita"
          },
          "observacoes": {
            "type": "string",
            "description": "Observações sobre a solicitação"
          },
          "dados_complementares": {
            "type": "object",
            "description": "Dados complementares específicos do tipo de benefício"
          },
          "documentos": {
            "description": "Documentos anexados à solicitação",
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/DocumentoSolicitacaoDto"
            }
          }
        },
        "required": [
          "beneficiario_id",
          "tipo_beneficio_id",
          "unidade_id",
          "documentos"
        ]
      },
...
    }
  }
}